+++ libs/common/src/dtos/find-query/find-query-where.dto.ts
@@ -1,6 +1,12 @@
 import { Singleton } from "@heap-code/singleton";
 import type { Type } from "@nestjs/common";
-import { Expose, plainToInstance, Transform, Type as TypeTransformer } from "class-transformer";
+import {
+	ClassTransformOptions,
+	Expose,
+	plainToInstance,
+	Transform,
+	Type as TypeTransformer
+} from "class-transformer";
 import {
 	IsArray,
 	isDateString,
@@ -11,6 +17,8 @@ import {
 } from "class-validator";
 
 import {
+	WhereBooleanDto,
+	WhereBooleanNullableDto,
 	WhereDateDto,
 	WhereDateNullableDto,
 	WhereNumberDto,
@@ -23,12 +31,20 @@ import { DtoPropertyOptions, dtoStorage, DtoType } from "../dto";
 
 export const UNKNOWN_DISCRIMINATED_TYPE = Symbol("unknown-type");
 
+/** @internal */
+interface Options {
+	dto?: DtoPropertyOptions<object>;
+	transformer?: ClassTransformOptions;
+}
+
 // Gets the "base" DTO for a given type
 /** @internal */
 function getWhereDtoType(type: DtoType, options: DtoPropertyOptions<object> = {}) {
 	const { nullable } = options;
 
 	switch (type) {
+		case Boolean:
+			return nullable ? WhereBooleanNullableDto : WhereBooleanDto;
 		case Date:
 			return nullable ? WhereDateNullableDto : WhereDateDto;
 		case Number:
@@ -45,7 +61,7 @@ function getWhereDtoType(type: DtoType, options: DtoPropertyOptions<object> = {}
 function transformWhereDto<T extends object>(
 	typeSingleton: Singleton<DtoType<T>>,
 	value: unknown,
-	options: DtoPropertyOptions<object> = {}
+	options: Options
 ): unknown {
 	if (value === undefined) {
 		return value;
@@ -57,9 +73,15 @@ function transformWhereDto<T extends object>(
 	// For "Primitive" types
 	switch (
 		!isObject(value) &&
-		(sourceType as unknown as typeof Date | typeof Number | typeof String)
+		(sourceType as unknown as typeof Boolean | typeof Date | typeof Number | typeof String)
 	) {
 		// Get back to this function with an object (for transformation)
+		case Boolean:
+			return transformWhereDto(
+				typeSingleton,
+				{ $eq: value as boolean } satisfies WhereBooleanDto,
+				options
+			);
 		case Date:
 			return transformWhereDto(
 				typeSingleton,
@@ -85,24 +107,28 @@ function transformWhereDto<T extends object>(
 			break;
 	}
 
-	const { discriminator } = options;
+	const { dto = {}, transformer } = options;
+	const { discriminator } = dto;
+
 	if (value !== null && discriminator) {
 		const { property, subTypes } = discriminator;
-		const discriminatedProperty = value[property];
+		const discriminatedProperty = isObject(value[property])
+			? (value[property] as WhereNumberDto | WhereStringDto).$eq
+			: value[property];
 		const propertyType = typeof discriminatedProperty;
 
 		// Only narrow type if the discriminator is "usable"
 		if (propertyType === "boolean" || propertyType === "number" || propertyType === "string") {
 			const subType = subTypes.find(({ name }) => name === discriminatedProperty);
 			if (subType) {
-				return plainToInstance(getWhereDtoType(subType.value), value);
+				return plainToInstance(getWhereDtoType(subType.value), value, transformer);
 			}
 
 			return UNKNOWN_DISCRIMINATED_TYPE;
 		}
 	}
 
-	return plainToInstance(getWhereDtoType(sourceType, options), value);
+	return plainToInstance(getWhereDtoType(sourceType, dto), value, transformer);
 }
 
 /** @internal */
@@ -111,14 +137,17 @@ function generateWhereType(source: DtoType, target: Type<unknown>) {
 		const type = new Singleton(() =>
 			dtoStorage.getPropertyType(source.prototype as Type<unknown>, key)
 		);
-		const options = dtoStorage.getPropertyOptions(source.prototype as Type<unknown>, key);
+		const dtoOptions = dtoStorage.getPropertyOptions(source.prototype as Type<unknown>, key);
 
 		Reflect.decorate(
 			[
 				Expose(),
 				IsOptional(),
-				Transform(({ key, obj }) =>
-					transformWhereDto(type, (obj as Record<string, unknown>)[key], options)
+				Transform(({ key, obj, options }) =>
+					transformWhereDto(type, (obj as Record<string, unknown>)[key], {
+						dto: dtoOptions,
+						transformer: options
+					})
 				),
 				NotEquals(UNKNOWN_DISCRIMINATED_TYPE, {
 					message: "The discriminated type was not determined"